app/requirements.txt
numpy==2.2.2
scikit-learn==1.6.1
fastapi==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.37
pydantic==2.10.6
pydantic-settings==2.7.1
sqlmodel==0.0.22
psycopg==3.2.4
psycopg-binary==3.2.4
bcrypt==4.2.1
email-validator==2.2.0
pyTelegramBotAPI==4.26.0
pocketsphinx==5.0.4
pyjwt==2.10.1
passlib==1.7.4
celery==5.4.0
pyamqp==0.1.0.7
Jinja2==3.1.5
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
psycopg==3.2.4
psycopg-binary==3.2.4
kombu==5.5.0rc3
transformers==4.49.0
torch===2.6.0
prometheus-client==0.21.1
pytest-env==1.1.5
Pillow==10.0
torchvision